--- zenovis/xinxinoptix/MaterialStuff.h
@@ -1,86 +1,0 @@
-#pragma once
-
-#include <cuda/helpers.h>
-#include "zxxglslvec.h"
-
-extern "C" {
-extern __constant__ Params params;
-}
-
-
-struct Onb
-{
-  __forceinline__ __device__ Onb(const vec3& normal)
-  {
-    m_normal = normal;
-
-    if( fabs(m_normal.x) > fabs(m_normal.z) )
-    {
-      m_binormal.x = -m_normal.y;
-      m_binormal.y =  m_normal.x;
-      m_binormal.z =  0;
-    }
-    else
-    {
-      m_binormal.x =  0;
-      m_binormal.y = -m_normal.z;
-      m_binormal.z =  m_normal.y;
-    }
-
-    m_binormal = normalize(m_binormal);
-    m_tangent = cross( m_binormal, m_normal );
-  }
-
-  __forceinline__ __device__ void inverse_transform(vec3& p) const
-  {
-    p = p.x*m_tangent + p.y*m_binormal + p.z*m_normal;
-  }
-
-  vec3 m_tangent;
-  vec3 m_binormal;
-  vec3 m_normal;
-};
-
-
-static __forceinline__ __device__ void setPayloadOcclusion( bool occluded )
-{
-    optixSetPayload_0( static_cast<unsigned int>( occluded ) );
-}
-
-
-static __forceinline__ __device__ void cosine_sample_hemisphere(const float u1, const float u2, vec3& p)
-{
-  // Uniformly sample disk.
-  const float r   = sqrtf( u1 );
-  const float phi = 2.0f*M_PIf * u2;
-  p.x = r * cosf( phi );
-  p.y = r * sinf( phi );
-
-  // Project up to hemisphere.
-  p.z = sqrtf( fmaxf( 0.0f, 1.0f - p.x*p.x - p.y*p.y ) );
-}
-
-#include "DisneyBRDF.h"
-
-struct MatOutput {
-    vec3  baseColor;
-    float metallic;
-    float roughness;
-    float subsurface;
-    float specular;
-    float specularTint;
-    float anisotropic;
-    float sheen;
-    float sheenTint;
-    float clearCoat;
-    float clearCoatGloss;
-    float opacity;
-};
-
-struct MatInput {
-    vec3 pos;
-    vec3 nrm;
-    vec3 uv;
-    vec3 clr;
-    vec3 tang;
-};
